+++ CHANGELOG.md
@@ -144,6 +144,8 @@ For instructions on upgrading to newer versions, visit
 
 ### Resolved Issues
 
+* \#2992 `Document.timeless` now properly scopes to the class and not thread.
+
 * \#2979 `pluck` no longer modifies the context in place. (Brian Goff)
 
 * \#2970 Fixed counter cache to properly use the name of the relation if available